requirements.txt
Flask>=2.3,<3.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-WTF>=1.1
pandas>=2.2
numpy>=2.0.0
dash>=2.11,<3.0
plotly>=5.17
python-dotenv>=1.0
gunicorn>=20.1
Flask-Mail>=0.9.1
itsdangerous>=2.1